refactor(hebrewLettersStore): split into letterSlice, audioSlice, drawingSlice, practiceSlice, statsSlice (#146)
