+++ gamesformykids/app/games/hebrew-letters/store/hebrewLettersStore.ts
@@ -1,418 +1,30 @@
-/**
- * ===============================================
- * Hebrew Letters Store — Zustand
- * ===============================================
- * מנהל את כל מצב משחק האותיות העבריות:
- * - האות הנוכחית + אותיות שהושלמו
- * - מצב אודיו
- * - מצב ציור (drawingState)
- * - מצב תרגול (practiceState)
- * - מצב עידוד (encouragementState)
- * - סטטיסטיקות למידה (learningStats)
- */
-
-import { create } from 'zustand';
+﻿import { create } from 'zustand';
 import { devtools } from 'zustand/middleware';
-import { TOTAL_HEBREW_LETTERS, type HebrewLetter } from '../constants/hebrewLetters';
-import {
-  DrawingState,
-  PracticeState,
-  EncouragementState,
-  LearningStats,
-} from '../types/hebrew-letters';
-import {
-  PRACTICE_STEPS,
-  DEFAULT_DRAWING_STATE,
-  DEFAULT_PRACTICE_STATE,
-  DEFAULT_ENCOURAGEMENT_STATE,
-  DEFAULT_AUDIO_STATE,
-  DEFAULT_LEARNING_STATS,
-  ENCOURAGEMENT_MESSAGES,
-  STEP_MESSAGES,
-} from '../constants/hebrewLettersConstants';
+import { TOTAL_HEBREW_LETTERS } from '../constants/hebrewLetters';
+import { PRACTICE_STEPS, STEP_MESSAGES } from '../constants/hebrewLettersConstants';
+import type { PracticeState } from '../types/hebrew-letters';
+import type { HebrewLetter } from '../constants/hebrewLetters';
+import type { HebrewLettersState, HebrewLettersStore } from './types';
 
-// ── Module-level refs ──────────────────────────────────────
-let encouragementTimerRef: ReturnType<typeof setTimeout> | null = null;
+import { createLetterSlice } from './slices/letterSlice';
+import { createAudioSlice } from './slices/audioSlice';
+import { createDrawingSlice } from './slices/drawingSlice';
+import { createPracticeSlice } from './slices/practiceSlice';
+import { createStatsSlice } from './slices/statsSlice';
 
-// ── State ──────────────────────────────────────────────────
-export interface HebrewLettersState {
-  currentLetter: HebrewLetter | null;
-  completedLetters: Set<string>;
-  isAudioEnabled: boolean;
-  drawingState: DrawingState;
-  practiceState: PracticeState;
-  encouragementState: EncouragementState;
-  learningStats: LearningStats;
-}
-
-// ── Actions ────────────────────────────────────────────────
-export interface HebrewLettersStoreActions {
-  // Letter management
-  setCurrentLetter: (letter: HebrewLetter | null) => void;
-  addCompletedLetter: (letterName: string) => void;
-  resetCompletedLetters: () => void;
-
-  // Audio
-  setIsAudioEnabled: (enabled: boolean) => void;
-  toggleAudio: () => void;
-  speakText: (text: string, settings?: { rate?: number; pitch?: number; volume?: number; lang?: string }) => void;
-  playLetterSound: (letter?: string) => void;
-  playEncouragementSound: () => void;
-  playStepCompletionSound: () => void;
-
-  // Drawing
-  updateDrawingState: (updates: Partial<DrawingState>) => void;
-  clearCanvas: () => void;
-  saveCanvasState: (imageData: ImageData) => void;
-  undoLastAction: () => void;
-  initializeCanvas: (width: number, height: number, backgroundColor?: string) => void;
-  startDrawing: (x: number, y: number) => void;
-  continueDrawing: (x: number, y: number) => void;
-  stopDrawing: () => void;
-  resetCanvas: () => void;
-  updateStrokeColor: (color: string) => void;
-  updateStrokeWidth: (width: number) => void;
-
-  // Practice
-  selectLetter: (letter: HebrewLetter) => void;
-  nextStep: () => void;
-  previousStep: () => void;
-  jumpToStep: (stepIndex: number) => void;
-  completeCurrentStep: () => void;
-  completeLetter: () => void;
-  resetPracticeProgress: () => void;
-
-  // Encouragement
-  showEncouragement: (message?: string, duration?: number) => void;
-  hideEncouragement: () => void;
-
-  // Session / stats
-  startPracticeSession: (letterName: string) => void;
-  endPracticeSession: () => void;
-  resetAllStats: () => void;
-}
+// Re-export interfaces for backward compat
+export type { HebrewLettersState, HebrewLettersStore };
+export type { HebrewLettersStoreActions } from './types';
 
 // ── Store ──────────────────────────────────────────────────
-export const useHebrewLettersStore = create<HebrewLettersState & HebrewLettersStoreActions>()(
+export const useHebrewLettersStore = create<HebrewLettersStore>()(
   devtools(
-    (set, get) => ({
-      // ── Initial state ──────────────────────────────────
-      currentLetter: null,
-      completedLetters: new Set<string>(),
-      isAudioEnabled: true,
-      drawingState: { ...DEFAULT_DRAWING_STATE },
-      practiceState: { ...DEFAULT_PRACTICE_STATE, completedSteps: new Set<number>() },
-      encouragementState: { ...DEFAULT_ENCOURAGEMENT_STATE },
-      learningStats: { ...DEFAULT_LEARNING_STATS },
-
-      // ── Letter management ──────────────────────────────
-      setCurrentLetter: (letter) =>
-        set({ currentLetter: letter }, false, 'hebrewLetters/setCurrentLetter'),
-
-      addCompletedLetter: (letterName) =>
-        set(
-          (s) => ({ completedLetters: new Set([...s.completedLetters, letterName]) }),
-          false,
-          'hebrewLetters/addCompletedLetter',
-        ),
-
-      resetCompletedLetters: () =>
-        set({ completedLetters: new Set<string>() }, false, 'hebrewLetters/resetCompleted'),
-
-      // ── Audio ──────────────────────────────────────────
-      setIsAudioEnabled: (enabled) =>
-        set({ isAudioEnabled: enabled }, false, 'hebrewLetters/setAudio'),
-
-      toggleAudio: () =>
-        set((s) => ({ isAudioEnabled: !s.isAudioEnabled }), false, 'hebrewLetters/toggleAudio'),
-
-      speakText: (text, settings) => {
-        const { isAudioEnabled } = get();
-        if (!isAudioEnabled || !('speechSynthesis' in window)) return;
-        const utterance = new SpeechSynthesisUtterance(text);
-        utterance.lang = settings?.lang ?? 'he-IL';
-        utterance.rate = settings?.rate ?? DEFAULT_AUDIO_STATE.speechRate;
-        utterance.pitch = settings?.pitch ?? DEFAULT_AUDIO_STATE.speechPitch;
-        utterance.volume = settings?.volume ?? DEFAULT_AUDIO_STATE.volume;
-        speechSynthesis.speak(utterance);
-      },
-
-      playLetterSound: (letter) => {
-        const { isAudioEnabled, currentLetter, speakText } = get();
-        if (!isAudioEnabled) return;
-        const letterToSpeak = letter ?? currentLetter?.letter ?? '';
-        const pronunciation = currentLetter?.pronunciation ?? letter ?? '';
-        speakText(`האות ${letterToSpeak}, ${pronunciation}`);
-      },
-
-      playEncouragementSound: () => {
-        const { isAudioEnabled, speakText } = get();
-        if (!isAudioEnabled) return;
-        const msg = ENCOURAGEMENT_MESSAGES[Math.floor(Math.random() * ENCOURAGEMENT_MESSAGES.length)];
-        speakText(msg);
-      },
-
-      playStepCompletionSound: () => {
-        const { isAudioEnabled, practiceState, speakText } = get();
-        if (!isAudioEnabled) return;
-        const msg = STEP_MESSAGES[practiceState.currentStep as keyof typeof STEP_MESSAGES];
-        if (msg) speakText(msg);
-      },
-
-      // ── Drawing ────────────────────────────────────────
-      updateDrawingState: (updates) =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, ...updates } }),
-          false,
-          'hebrewLetters/updateDrawing',
-        ),
-
-      clearCanvas: () =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, paths: [], isDrawing: false } }),
-          false,
-          'hebrewLetters/clearCanvas',
-        ),
-
-      saveCanvasState: (imageData) =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, paths: [...s.drawingState.paths, imageData] } }),
-          false,
-          'hebrewLetters/saveCanvas',
-        ),
-
-      undoLastAction: () =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, paths: s.drawingState.paths.slice(0, -1) } }),
-          false,
-          'hebrewLetters/undo',
-        ),
-
-      initializeCanvas: (width, height, backgroundColor = '#ffffff') =>
-        set(
-          (s) => ({
-            drawingState: {
-              ...s.drawingState,
-              canvasWidth: width,
-              canvasHeight: height,
-              backgroundColor,
-              paths: [],
-              isDrawing: false,
-              lastDrawPosition: null,
-            },
-          }),
-          false,
-          'hebrewLetters/initCanvas',
-        ),
-
-      startDrawing: (x, y) =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, isDrawing: true, lastDrawPosition: { x, y } } }),
-          false,
-          'hebrewLetters/startDrawing',
-        ),
-
-      continueDrawing: (x, y) => {
-        // Read live state to avoid stale closure
-        if (!get().drawingState.isDrawing) return;
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, lastDrawPosition: { x, y } } }),
-          false,
-          'hebrewLetters/continueDrawing',
-        );
-      },
-
-      stopDrawing: () =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, isDrawing: false, lastDrawPosition: null } }),
-          false,
-          'hebrewLetters/stopDrawing',
-        ),
-
-      resetCanvas: () =>
-        set(
-          { drawingState: { ...DEFAULT_DRAWING_STATE } },
-          false,
-          'hebrewLetters/resetCanvas',
-        ),
-
-      updateStrokeColor: (color) =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, currentStrokeColor: color } }),
-          false,
-          'hebrewLetters/strokeColor',
-        ),
-
-      updateStrokeWidth: (width) =>
-        set(
-          (s) => ({ drawingState: { ...s.drawingState, currentStrokeWidth: width } }),
-          false,
-          'hebrewLetters/strokeWidth',
-        ),
-
-      // ── Practice ───────────────────────────────────────
-      selectLetter: (letter) =>
-        set(
-          {
-            currentLetter: letter,
-            practiceState: { ...DEFAULT_PRACTICE_STATE, completedSteps: new Set<number>() },
-            encouragementState: { ...DEFAULT_ENCOURAGEMENT_STATE },
-          },
-          false,
-          'hebrewLetters/selectLetter',
-        ),
-
-      nextStep: () =>
-        set(
-          (s) => {
-            const next = Math.min(s.practiceState.currentStep + 1, PRACTICE_STEPS.length - 1);
-            return {
-              practiceState: {
-                ...s.practiceState,
-                currentStep: next,
-                completedSteps: new Set([...s.practiceState.completedSteps, s.practiceState.currentStep]),
-              },
-            };
-          },
-          false,
-          'hebrewLetters/nextStep',
-        ),
-
-      previousStep: () =>
-        set(
-          (s) => ({
-            practiceState: {
-              ...s.practiceState,
-              currentStep: Math.max(s.practiceState.currentStep - 1, 0),
-            },
-          }),
-          false,
-          'hebrewLetters/prevStep',
-        ),
-
-      jumpToStep: (stepIndex) => {
-        if (stepIndex < 0 || stepIndex >= PRACTICE_STEPS.length) return;
-        set(
-          (s) => ({ practiceState: { ...s.practiceState, currentStep: stepIndex } }),
-          false,
-          'hebrewLetters/jumpToStep',
-        );
-      },
-
-      completeCurrentStep: () => {
-        const { practiceState, showEncouragement } = get();
-        if (practiceState.completedSteps.has(practiceState.currentStep)) return;
-        set(
-          (s) => ({
-            practiceState: {
-              ...s.practiceState,
-              completedSteps: new Set([...s.practiceState.completedSteps, s.practiceState.currentStep]),
-            },
-          }),
-          false,
-          'hebrewLetters/completeStep',
-        );
-        showEncouragement();
-      },
-
-      completeLetter: () => {
-        const { currentLetter, learningStats, showEncouragement } = get();
-        if (!currentLetter) return;
-        const letterName = currentLetter.letter;
-        const sessionTime = learningStats.sessionStartTime
-          ? Date.now() - learningStats.sessionStartTime
-          : 0;
-        set(
-          (s) => ({
-            completedLetters: new Set([...s.completedLetters, letterName]),
-            practiceState: { ...s.practiceState, isCompleted: true } as PracticeState,
-            learningStats: {
-              ...s.learningStats,
-              lettersCompleted: new Set([...s.learningStats.lettersCompleted, letterName]),
-              totalPracticeTime: s.learningStats.totalPracticeTime + sessionTime,
-              sessionStartTime: 0,
-              practiceHistory: [
-                ...s.learningStats.practiceHistory,
-                {
-                  letter: letterName,
-                  stepCompleted: s.practiceState.completedSteps.size,
-                  timeSpent: sessionTime,
-                  timestamp: Date.now(),
-                },
-              ],
-            },
-          }),
-          false,
-          'hebrewLetters/completeLetter',
-        );
-        showEncouragement('כל הכבוד! סיימת את התרגול בהצלחה!', 3000);
-      },
-
-      resetPracticeProgress: () =>
-        set(
-          { practiceState: { ...DEFAULT_PRACTICE_STATE, completedSteps: new Set<number>() } },
-          false,
-          'hebrewLetters/resetPractice',
-        ),
-
-      // ── Encouragement ──────────────────────────────────
-      showEncouragement: (message, duration = 2000) => {
-        if (encouragementTimerRef) clearTimeout(encouragementTimerRef);
-        const msg = message ?? ENCOURAGEMENT_MESSAGES[Math.floor(Math.random() * ENCOURAGEMENT_MESSAGES.length)];
-        set(
-          { encouragementState: { showEncouragement: true, currentMessage: msg, isStepCompleted: false } },
-          false,
-          'hebrewLetters/showEncouragement',
-        );
-        encouragementTimerRef = setTimeout(() => {
-          set(
-            (s) => ({ encouragementState: { ...s.encouragementState, showEncouragement: false } }),
-            false,
-            'hebrewLetters/hideEncouragement',
-          );
-          encouragementTimerRef = null;
-        }, duration);
-      },
-
-      hideEncouragement: () => {
-        if (encouragementTimerRef) { clearTimeout(encouragementTimerRef); encouragementTimerRef = null; }
-        set(
-          (s) => ({ encouragementState: { ...s.encouragementState, showEncouragement: false } }),
-          false,
-          'hebrewLetters/hideEncouragement',
-        );
-      },
-
-      // ── Session / stats ────────────────────────────────
-      startPracticeSession: (letterName) =>
-        set(
-          (s) => ({
-            learningStats: {
-              ...s.learningStats,
-              lettersStarted: new Set([...s.learningStats.lettersStarted, letterName]),
-              sessionStartTime: Date.now(),
-            },
-          }),
-          false,
-          'hebrewLetters/startSession',
-        ),
-
-      endPracticeSession: () =>
-        set(
-          (s) => ({ learningStats: { ...s.learningStats, sessionStartTime: 0 } }),
-          false,
-          'hebrewLetters/endSession',
-        ),
-
-      resetAllStats: () =>
-        set(
-          {
-            completedLetters: new Set<string>(),
-            learningStats: { ...DEFAULT_LEARNING_STATS },
-          },
-          false,
-          'hebrewLetters/resetAllStats',
-        ),
+    (...a) => ({
+      ...createLetterSlice(...a),
+      ...createAudioSlice(...a),
+      ...createDrawingSlice(...a),
+      ...createPracticeSlice(...a),
+      ...createStatsSlice(...a),
     }),
     { name: 'hebrewLettersStore' },
   ),
